refactor: split utils to multiple files #208
