+++ src/utils/notNullOrUndefined.js
@@ -1,0 +1,3 @@
+export function notNullOrUndefined(value) {
+  return value !== null && value !== void 0;
+}
