+++ src/utils/rootLocation.js
@@ -1,0 +1,3 @@
+export function rootLocation() {
+  return { start: { line: 0, column: 0 }, end: { line: 0, column: 0 } };
+}
